# Optimize CPU and Memory performance for Resize linear mode parser