Re-write calc_neighbor_points() by composing index from binary bits instead of recursion.

With the optimized calc_neighbor_points(), CPU time required by 90% and peak memory utilization is significantly reduced.

Perf. comparision on VM w/ 12-Core EPYC 9V64 + 128 GB mem:

| n_dim | out_elements | New t-CPU (us) | Old t-CPU (us) | t-CPU Ratio |
|------:|-------------:|---------------:|---------------:|------------:|
|     4 |       786432 |         120405 | 1494350        | 0.0806      |
|     4 |      1572864 |         282763 | 3826060        | 0.0739      |
|     4 |      3145728 |         650957 | 7941436        | 0.0820      |
|     4 |      6291456 |        1304652 | 14869059       | 0.0877      |
|     4 |     12582912 |        2608523 | 29432326       | 0.0886      |
|     4 |     25165824 |        5175560 | 58848631       | 0.0879      |
|     4 |     50331648 |       10486676 | 118005802      | 0.0889      |
|     4 |    100663296 |       21141464 | OOM Kill       | N/A         |