@coxuamd 2024-12-24
> Thank you for taking a stab at cleaning this function. However, this PR needs two additional things (besides all test-cases to pass):
> 
> * Code comments for `calc_neighbor_points`. And this is not the fault of this PR, but the comments were/are entirely missing, yet this function is very complicated for a reviewer's understanding, and I am not sure what exactly is it doing. And that documentation needs to be fixed now, since this function is being rewritten.
> * A unit-test to specifically test this extremely complex function. It should have been there much earlier, but it can be added now in this PR.

Here's the explanation of the new algorithm:
Given a std::vector<std::vector<std::vector<std::size_t>>> vvv_ind() in below layout:
vvv_ind{} = {
    {{}, {}},
    {{}, {}},
    {{}, {}},
    {{}, {}},
}

1-dim is n_dims, example n_dims = 4;
2-dim is hard-coded to 2, in the original algorithm, ref. as lo/hi
3-dim is a vector of integer with out_elements.

What the original calc_neighbor_points() algorithm is trying to do, is to compose a new vector, with size 2^n_dims, within each is a n_dim vector of integer. In current example, if out_elements=16, n_dims=4, then we'll have a new vector as below:

vec_ind{} = {
    {}, # Each element is a vector of 4 integer
    {},
    {},
    ..
    {}, # 2^4 * 16 vectors in total
}

Let's re-write vvv_ind in different pattern for friendly understanding, where each character is a vector of 16 elements, like A = {0,1,1,1,0,1,1,1,0,1,1,1,0,1,1,1}

AB
CD
EF
GH

The original calc_neighbor_points() is done in a recursive way that append elements in vertical, and expand in horizontal.

Recursion:
#0:
A
B

#1:
AC
BC
AD
BD

#2:
ACE
BCE
ADE
BDE
ACF
BCF
ADF
BDF

#3:
ACEG
BCEG
ADEG
BDEG
...
BDFG

#4, last time, with the crafted vector, 2^4 * 16 elements, each element is a vector of 4 integer as the index, to get the index from shape in_s.index(idx)

Since the 2nd dimension is hardcoded to 2, we can treat this dimension as binary. What the final result (before in_s.index(idx)) is actaully we increase from 0 to 2^n_dim, and convert this n-dim bits value to binary, using the bit to index into 2nd-dim of vvv_ind (hi or low), use the position in n-dim to index into 1st-dim of vvv_ind, and loop out_elements so that all elements in A (and other capital character) can be looped.

Taking #3 to explain:

vec_ind{}
0-1
AB
CD
EF
GH

#3:
outer loop on 2^n_dims w/ start:
ACEG >> 0000b >> start(0) convert to 4-bit binary, using each bit to index to vec_ind
  middle loop on out_elements, so that we process all out_elements(16) in same way.
    reset bits shift from start, so on each element, shift can cover all bits.
    inner loop: Pick each bit to index to vec_ind
BCEG >> 1000b >> start(1), ditto
ADEG >> 0100b >> start(2), ditto
BDEG >> 1100b >> start(3), ditto
...
BDFG >> 1111b